Add SideNavItems component, also move all SideNav-related components to SideNav folder
