+++ src/components/SideNav/SideNavIcon.js
